jupyterlab-native
jupyterlab-native is a build-time dependency for jupyter-notebook (and/or its
components). It is mostly an exercise is allowing numerous python recipes (and
their dependants, and their dependants, etc) to be built natively. This patch
is submitted separately to clearly delineate between that which is required
for jupyter-lab versus that which is required for jupyter-notebook without
muddying those commits with native handling.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>