jupyter-notebook: add/update to scarthgap
Add/update all recipes to create a working jupyter-notebook for scarthgap. If
a required recipe already exists in one of the upstream required layers
(oe-core and meta-oe/meta-python) then use that recipe even if it is not the
absolute latest. For all other required recipes, use the absolute latest code
as published on pypi.org.

Updated:
	python3-jupyter-notebook: 6.4.12 → 7.2.1

New:
	python3-argon2-cffi-bindings: 21.2.0
	python3-fqdn: 1.5.1
	python3-isoduration: 20.11.0
	python3-uri-template: 1.3.0

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>